+++ test/unit/fuzz-tools.test.js
@@ -491,6 +491,7 @@ import * as asrsV513 from '../../lib/asrs-v513.js';
 import * as ymrsV514 from '../../lib/ymrs-v514.js';
 import * as simpsonAngusV515 from '../../lib/simpson-angus-v515.js';
 import * as asthmaControlTestV516 from '../../lib/asthma-control-test-v516.js';
+import * as pippV517 from '../../lib/pipp-v517.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -927,6 +928,7 @@ const MODULES = {
   'ymrs-v514.js': ymrsV514,
   'simpson-angus-v515.js': simpsonAngusV515,
   'asthma-control-test-v516.js': asthmaControlTestV516,
+  'pipp-v517.js': pippV517,
   'tb-testing.js': tbTesting,
 };